[ENG-7966] Add "collected-in" relationship for Nodes (#11140)
## Purpose
adds new relationship for collections a node is in.

## Changes
- splits up node detail tests into more workable files <500 that are more atomic
- add related view for collection to nodes detail
- adds tests 

## Ticket
https://openscience.atlassian.net/browse/ENG-7966
